[sigcomm-demo-do-not-touch] dhcp and arp default rules get higher priority
this is to fix an issue with macs: because they try to reuse their previous
dhcp address, you cannot rely on a dhcp being the first thing you see from
a host that has associated- this typically means that an overlap rule
will get inserted that means that the real dhcp packet does *not* get
forwarded to the controller, so isn't processed, so the client gets confused
and everything breaks. making the default dhcp rule higher priority should
mean that dhcp really is always processed by the controller and hence the
server so all is well.